commit changes

+++ src/KingHotelProject.Application/DTOs/DishDtos.cs
@@ -4,14 +4,14 @@
     {
         public Guid DishId { get; set; }
         public string DishName { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public Guid HotelId { get; set; }
     }
 
     public class DishCreateDto
     {
         public string DishName { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public Guid HotelId { get; set; }
     }
 
@@ -23,6 +23,6 @@
     public class DishUpdateDto
     {
         public string DishName { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
